#184 refactor: 테스트에 fixture 패키지 생성
