+++ src/test/java/com/floney/floney/book/MonthLineResponseTest.java
@@ -2,12 +2,12 @@ package com.floney.floney.book;
 
 import com.floney.floney.book.dto.process.BookLineExpense;
 import com.floney.floney.book.dto.response.MonthLinesResponse;
+import com.floney.floney.fixture.BookLineFixture;
+import java.util.Arrays;
+import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import java.util.Arrays;
-import java.util.List;
 
 public class MonthLineResponseTest {
