+++ src/test/java/com/floney/floney/config/WithMockCustomUser.java
@@ -1,12 +1,7 @@
 package com.floney.floney.config;
 
-import com.floney.floney.config.WithMockCustomUserSecurityContextFactory;
-import com.floney.floney.user.dto.constant.Role;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
-import java.util.Collection;
-import java.util.Set;
-import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.test.context.support.TestExecutionEvent;
 import org.springframework.security.test.context.support.WithSecurityContext;
